requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-PyMongo==2.3.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
PyJWT==1.4.2
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
regex==2021.4.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
